corigindo bug no delet

+++ src/main/java/com/example/sicapweb/web/controller/ap/concurso/DocumentoAdmissaoController.java
@@ -177,7 +177,7 @@ public class DocumentoAdmissaoController extends DefaultController<DocumentoAdmi
     @DeleteMapping(value = {"/{id}"})
     public void delete(@PathVariable BigInteger id) {
         DocumentoAdmissao documentoAdmissao = documentoAdmissaoRepository.findById(id);
-        if(documentoAdmissao.getOpcaoDesistencia() == 5){
+        if(documentoAdmissao.getOpcaoDesistencia() != null && documentoAdmissao.getOpcaoDesistencia() == 5 ){
             editalAprovadoRepository.delete(documentoAdmissao.getEditalAprovado().getId());
         }
         if (documentoAdmissao.getDocumentoCastorId()!=null ) throw new RuntimeException("remova  primeiro o documento antes de excluir a aprovado!!");
